wtapctl: fix build errors in userland tool
if_wtapioctl.h: add MAX_NBR_WTAP and WTAP_HAL_BITMAP_SIZE definitions.
These constants were only in if_wtapvar.h (kernel-only), but wtapctl.c
needs them for the device_list() bitmap array.  Moving them to the
ioctl header (which is shared between kernel and userland) makes them
available to the userland tool without pulling in kernel-only headers.

wtapctl.c: cast VIS_MAP_NWORDS to int in the vis_link_show() loop to
silence -Wsign-compare/-Werror on the int/size_t comparison.

+++ usr.sbin/wtapctl/wtapctl.c
@@ -140,7 +140,7 @@ vis_link_show(const struct cmd *cmd __unused, int id)
 	printf("medium: %s\n", is_open ? "open" : "closed");
 	printf("wtap%d ->", id);
 
-	for (int i = 0; i < VIS_MAP_NWORDS; i++) {
+	for (int i = 0; i < (int)VIS_MAP_NWORDS; i++) {
 		uint32_t word = req.map[i];
 
 		for (int j = 0; j < bpw; j++) {
